added updating status and points

+++ backend/src/main/java/com/redlions/backend/entity/Profile.java
@@ -37,7 +37,7 @@ public class Profile {
     private String password;
 
     @Column(name="points")
-    private Long points;
+    private Integer points;
 
     @Column(name="happiness")
     private Long happiness;
@@ -79,7 +79,7 @@ public class Profile {
 
     }
 
-    public Profile(Long id, String name, String email, String password, Long points, Long happiness, String aboutMe, String profilePicture, Float busyness) {
+    public Profile(Long id, String name, String email, String password, Integer points, Long happiness, String aboutMe, String profilePicture, Float busyness) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -123,11 +123,11 @@ public class Profile {
         this.password = password;
     }
 
-    public Long getPoints() {
+    public Integer getPoints() {
         return this.points;
     }
 
-    public void setPoints(Long points) {
+    public void setPoints(Integer points) {
         this.points = points;
     }
